Grocery list

+++ backend/src/main/java/com/uo/price_comparator/dto/OfferDto.java
@@ -1,24 +1,25 @@
 package com.uo.price_comparator.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class OfferDto {
     private String store;
-    private double price;
-    private double pricePerUnit;
+    private BigDecimal price;
+    private BigDecimal pricePerUnit;
     private LocalDate date;
     private Integer discountPercent;
-    private Double discountedPrice;
-    private Double discountedPricePerUnit;
+    private BigDecimal discountedPrice;
+    private BigDecimal discountedPricePerUnit;
 
     public String getStore() { return store; }
     public void setStore(String store) { this.store = store; }
 
-    public double getPrice() { return price; }
-    public void setPrice(double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
-    public double getPricePerUnit() { return pricePerUnit; }
-    public void setPricePerUnit(double pricePerUnit) { this.pricePerUnit = pricePerUnit; }
+    public BigDecimal getPricePerUnit() { return pricePerUnit; }
+    public void setPricePerUnit(BigDecimal pricePerUnit) { this.pricePerUnit = pricePerUnit; }
 
     public LocalDate getDate() { return date; }
     public void setDate(LocalDate date) { this.date = date; }
@@ -26,9 +27,9 @@ public class OfferDto {
     public Integer getDiscountPercent() { return discountPercent; }
     public void setDiscountPercent(Integer discountPercent) { this.discountPercent = discountPercent; }
 
-    public Double getDiscountedPricePerUnit() { return discountedPricePerUnit; }
-    public void setDiscountedPricePerUnit(Double discountedPricePerUnit) { this.discountedPricePerUnit = discountedPricePerUnit; }
+    public BigDecimal getDiscountedPricePerUnit() { return discountedPricePerUnit; }
+    public void setDiscountedPricePerUnit(BigDecimal discountedPricePerUnit) { this.discountedPricePerUnit = discountedPricePerUnit; }
 
-    public Double getDiscountedPrice() { return discountedPrice; }
-    public void setDiscountedPrice(Double discountedPrice) { this.discountedPrice = discountedPrice; }
+    public BigDecimal getDiscountedPrice() { return discountedPrice; }
+    public void setDiscountedPrice(BigDecimal discountedPrice) { this.discountedPrice = discountedPrice; }
 }
